Organised code for donation menu patches
