+++ stardew-access/Patches/ChatMenuPatches.cs
@@ -19,23 +19,22 @@ namespace stardew_access.Patches
                     bool isPrevButtonPressed = MainClass.Config.ChatMenuNextKey.JustPressed();
                     bool isNextButtonPressed = MainClass.Config.ChatMenuPreviousKey.JustPressed();
 
-                    if (___messages.Count > 0)
+                    if (___messages.Count <= 0) return;
+
+                    #region To narrate previous and next chat messages
+                    if (isNextButtonPressed && !isChatRunning)
                     {
-                        #region To narrate previous and next chat messages
-                        if (isNextButtonPressed && !isChatRunning)
-                        {
-                            isChatRunning = true;
-                            CycleThroughChatMessages(true, ___messages);
-                            Task.Delay(200).ContinueWith(_ => { isChatRunning = false; });
-                        }
-                        else if (isPrevButtonPressed && !isChatRunning)
-                        {
-                            isChatRunning = true;
-                            CycleThroughChatMessages(false, ___messages);
-                            Task.Delay(200).ContinueWith(_ => { isChatRunning = false; });
-                        }
-                        #endregion
+                        isChatRunning = true;
+                        CycleThroughChatMessages(true, ___messages);
+                        Task.Delay(200).ContinueWith(_ => { isChatRunning = false; });
                     }
+                    else if (isPrevButtonPressed && !isChatRunning)
+                    {
+                        isChatRunning = true;
+                        CycleThroughChatMessages(false, ___messages);
+                        Task.Delay(200).ContinueWith(_ => { isChatRunning = false; });
+                    }
+                    #endregion
                 }
                 else if (___messages.Count > 0)
                 {
